pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "bcrypt==4.3.0",
    "certifi==2025.1.31",
    "click==8.1.8",
    "colorama==0.4.6",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "exceptiongroup==1.2.2",
    "fastapi-cli[standard]==0.0.7",
    "fastapi[standard]==0.115.12",
    "graphviz>=0.20.3",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "httpcore==1.0.8",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "passlib[bcrypt]==1.7.4",
    "psycopg-binary==3.2.6",
    "psycopg-pool==3.2.6",
    "psycopg[binary,pool]==3.2.6",
    "pydantic-core==2.23.4",
    "pydantic==2.9.2",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "rich-toolkit==0.14.1",
    "rich==14.0.0",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "starlette==0.46.1",
    "typer==0.15.2",
    "typing-extensions==4.13.2",
    "tzdata==2025.2",
    "uvicorn[standard]==0.34.0",
    "watchfiles==1.0.5",
    "websockets==15.0.1",
]
